// src/models/knack.model.ts
export interface KnackApplication {
    _id: string
    users: Users
    ecommerce: Ecommerce
    counts: Counts
    field_count: number
    thumb_count: number
    object_count: number
    task_count: number
    view_count: number
    scene_count: number
    credential_count: number
    status: string
    settings: Settings
    tasks: any[]
    payment_processors: any[]
    design: Design
    layout: Layout
    copying: boolean
    feature_flags: boolean[]
    name: string
    slug: string
    distributions: any[]
    versions: Version[]
    first_created: string
    account_id: string
    user_id: string
}

export interface Users {
    enabled: boolean
    scope: string
    registration: string
}

export interface Ecommerce {
    enabled: boolean
}

export interface Counts {
    total_entries: number
    asset_count: number
    asset_size: number
}

export interface Settings {
    geo: boolean
    timezone: string
    timezone_offset: string
    timezone_dst: string
    cluster: string
    powered_by_link: boolean
    new_count: boolean
    registration_migration_status: string
    https_redirect: boolean
    should_purge_record_history: boolean
    support_access: boolean
    use_multiple_api_subdomains: boolean
    scriptProtectionEnabled: boolean
    embed_login_method: string
    ignoreBilling: boolean
    v3_beta: boolean
    v3_open_beta: boolean
    mongo: string
    solr: string
}

export interface Design {
    general: General
    regions: Regions
    components: Components
}

export interface General {
    theme: string
    font: string
    links: Links
    buttons: Buttons
    tables: Tables
}

export interface Links {
    color: string
}

export interface Buttons {
    color: string
    bg_color: string
}

export interface Tables {
    style: string
    dividers: boolean
    striped: boolean
    hover: boolean
    spacing: string
}

export interface Regions {
    body: Body
    header: Header
}

export interface Body {
    full_width: boolean
}

export interface Header {
    bg_color: string
    menu: Menu
    title: Title
}

export interface Menu {
    show: boolean
    user_auth_based: boolean
    format: string
    color: string
    outline_or_fill_color: string
}

export interface Title {
    color: string
    show_logo: boolean
}

export interface Components {
    notifications: Notifications
}

export interface Notifications {
    color: string
    bg_color: string
}

export interface Layout {
    entry_scene_menu: boolean
    app_menu_auth: boolean
    theme: string
}

export interface Version {
    _id: string
    status: string
    objects: Object[]
    scenes: Scene[]
}

export interface Object {
    inflections: Inflections
    connections: Connections
    user: boolean
    status: string
    tasks: any[]
    type: string
    _id: string
    name: string
    fields: Field[]
    template?: string
    key: string
    identifier: string
    profile_key?: string
}
// ...
export interface Scene {
    groups: Group[]
    _id: string
    name: string
    slug: string
    key: string
    views: View[]
    parent?: string
    object?: string
    type?: string
    limit_profile_access?: boolean
    allowed_profiles?: any[]
}